Fix: Client 쪽과 맞지 않았던 부분 수정 #176
