+++ src/main/java/com/nhnacademy/marketgg/server/controller/PostController.java
@@ -123,7 +123,7 @@ public class PostController {
     /**
      * 회원의 1:1 문의 목록을 조회하는 GET Mapping 을 지원합니다.
      *
-     * @param page - 페이징 처리를 위한 페이지 번호입니다.
+     * @param page       - 페이징 처리를 위한 페이지 번호입니다.
      * @param memberInfo - 목록을 조회하는 회원의 정보가 담긴 객체입니다.
      * @return 조회한 1:1 문의 목록을 List 로 반환합니다.
      * @since 1.0.0
@@ -158,58 +158,6 @@ public class PostController {
 
         return ResponseEntity.status(HttpStatus.OK)
                              .location(URI.create(DEFAULT_POST + "/categories/" + categoryCode + "/search"))
-                             .body(responses);
-    }
-
-    /**
-     * 지정한 게시판 타입의 Reason 옵션으로 검색한 결과를 반환합니다.
-     *
-     * @param categoryCode  - 검색을 진행 할 게시판 타입입니다.
-     * @param option        - 검색을 진행 할 필터의 값입니다.
-     * @param searchRequest - 검색을 진행 할 검색정보입니다.
-     * @return 검색정보로 검색한 결과 목록 응답객체를 반환합니다.
-     * @throws ParseException          파싱도중 예외처리입니다.
-     * @throws JsonProcessingException JSON 관련 파싱처리 도중 예외처리입니다.
-     * @since 1.0.0
-     */
-    @PostMapping("/categories/{categoryCode}/search/reason")
-    public ResponseEntity<List<SearchBoardResponse>> searchPostListForReason(@PathVariable final String categoryCode,
-                                                                             @RequestParam final String option,
-                                                                             @RequestBody final SearchRequest searchRequest)
-            throws ParseException, JsonProcessingException {
-
-        List<SearchBoardResponse> responses =
-                postService.searchForOption(categoryCode, searchRequest, option, "reason");
-
-        return ResponseEntity.status(HttpStatus.OK)
-                             .location(URI.create(
-                                     DEFAULT_POST + "/categories/" + categoryCode + "/search/reason/" + option))
-                             .body(responses);
-    }
-
-    /**
-     * 지정한 게시판 타입의 Status 옵션으로 검색한 결과를 반환합니다.
-     *
-     * @param categoryCode  - 검색을 진행 할 게시판 타입입니다.
-     * @param option        - 검색을 진행 할 필터의 값입니다.
-     * @param searchRequest - 검색을 진행 할 검색정보입니다.
-     * @return 검색정보로 검색한 결과 목록 응답객체를 반환합니다.
-     * @throws ParseException          파싱도중 예외처리입니다.
-     * @throws JsonProcessingException JSON 관련 파싱처리 도중 예외처리입니다.
-     * @since 1.0.0
-     */
-    @PostMapping("/categories/{categoryCode}/search/status")
-    public ResponseEntity<List<SearchBoardResponse>> searchPostListForStatus(@PathVariable final String categoryCode,
-                                                                             @RequestParam final String option,
-                                                                             @RequestBody final SearchRequest searchRequest)
-            throws ParseException, JsonProcessingException {
-
-        List<SearchBoardResponse> responses =
-                postService.searchForOption(categoryCode, searchRequest, option, "status");
-
-        return ResponseEntity.status(HttpStatus.OK)
-                             .location(URI.create(
-                                     DEFAULT_POST + "/categories/" + categoryCode + "/search/status/" + option))
                              .body(responses);
     }
